fix(env-utils): move publicKey from `@suite-common/wallet-constants`
