+++ packages/env-utils/tsconfig.json
@@ -1,9 +1,5 @@
 {
     "extends": "../../tsconfig.base.json",
     "compilerOptions": { "outDir": "libDev" },
-    "references": [
-        {
-            "path": "../../suite-common/wallet-constants"
-        }
-    ]
+    "references": []
 }
