+++ suite-common/wallet-constants/src/index.ts
@@ -3,4 +3,3 @@ export * from './formDraft';
 export * from './polling';
 export * from './sendForm';
 export * from './discovery';
-export * from './jws';
